apirest-aplication/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-keycloak
pydantic
keycloak-client
